Move the SSO token API into another interface
