+++ common4j/src/main/com/microsoft/identity/common/java/AuthenticationConstants.java
@@ -546,6 +546,7 @@ public class AuthenticationConstants {
          * String of broker client ID.
          */
         public static final String BROKER_CLIENT_ID = "29d9ed98-a469-4536-ade2-f981bc1d605e";
+
     }
 
     @NoArgsConstructor(access = AccessLevel.PRIVATE)
